+++ appFrontend/src/App.tsx
@@ -11,13 +11,7 @@ import { fetch } from '@react-native-community/netinfo';
 import { Alert } from 'react-native';
 import RNExitApp from 'react-native-exit-app';
 import { RootStackParamList } from './navigation/types/navigation';
-import messaging, { FirebaseMessagingTypes } from '@react-native-firebase/messaging';
-import notifee, {
-  AndroidImportance,
-  AndroidVisibility,
-  EventDetail,
-  EventType,
-} from '@notifee/react-native';
+import notifee from '@notifee/react-native';
 import { Walkthrough } from './screens/homeScreen/components';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 
@@ -91,41 +85,37 @@ const App = (): JSX.Element => {
     }
   }, [isWalkthroughClosed]);
 
-  // 포그라운드 알림 수신
-  useEffect(() => {
-    const unsubscribe = messaging().onMessage(async (remoteMessage) => {
-      onMessageReceived(remoteMessage);
-    });
-    return unsubscribe;
-  }, []);
+  // const handleNotificationPress = (remoteMessage) => {
+  //   const pathJsonString = remoteMessage.notification?.data?.path;
+  //   (pathJsonString: string) => {
+  //     try {
+  //       console.log('알림 클릭: ', pathJsonString);
+  //       return JSON.parse(pathJsonString);
+  //     } catch (error) {
+  //       console.error('JSON 파싱 오류:', error);
+  //     }
+  //   };
+  // };
 
-  // 백그라운드 알림 수신 및 표시
-  messaging().setBackgroundMessageHandler(async (remoteMessage) => {});
+  // useEffect(() => {
+  //   // foreground일때 notifee 알림제어
+  //   notifee.onForegroundEvent(async ({ type, detail }) => {
+  //     if (type === EventType.PRESS) {
+  //       const pathJsonString = detail.notification?.data?.path as string;
+  //       console.log('detail', detail);
+  //       const aaa = await JSON.parse(pathJsonString);
+  //       await console.log('파싱값aaa:', aaa);
+  //     }
+  //   });
 
-  // 포그라운드 알림 화면에 표시
-  const onMessageReceived = async (message: FirebaseMessagingTypes.RemoteMessage) => {
-    const { title, body } = message.notification!;
-
-    // 채널 생성
-    const channelId = await notifee.createChannel({
-      // Miscellaneous 채널 삭제, Important Notifications로 통임
-      id: 'fcm_fallback_notification_channel',
-      name: 'Important Notifications',
-      importance: AndroidImportance.HIGH,
-    });
-
-    // 디바이스에 알림 표시
-    await notifee.displayNotification({
-      title,
-      body,
-      android: {
-        channelId,
-        importance: AndroidImportance.HIGH,
-        visibility: AndroidVisibility.PUBLIC,
-        timeoutAfter: 2000,
-      },
-    });
-  };
+  //   // background, quit일떄 알림제어 -> ios는 quit상태가 없어서 필요없지만 더 공부해봄
+  //   notifee.onBackgroundEvent(async ({ type, detail }) => {
+  //     if (type === EventType.PRESS) {
+  //       const pathJsonString = detail.notification?.data?.path as string;
+  //       JSON.parse(pathJsonString);
+  //     }
+  //   });
+  // }, []);
 
   return (
     <Provider store={store}>
