+++ appFrontend/src/navigation/HomeNavigation.tsx
@@ -12,6 +12,7 @@ import NotiHistoryScreen from '@/screens/homeScreen/components/NotiHistory';
 import HomeScreen from '@/screens/homeScreen';
 import SavedRoutesScreen from '@/screens/savedRoutesScreen';
 import toastConfig from '@global/utils/ToastConfig';
+import { pushNotification } from '@/global/utils/pushNotification';
 
 const Stack = createStackNavigator<HomeStackParamList>();
 
@@ -20,6 +21,7 @@ const screenOption = {
 };
 
 const HomeNavigation = () => {
+  pushNotification();
   return (
     <>
       <Stack.Navigator initialRouteName={HOME} screenOptions={screenOption}>
